pythia/templates/pipeline-worker/{{cookiecutter.worker_name}}/requirements.txt
# Core Pythia dependencies
pythia>=0.1.0

# Message broker dependencies
{% if cookiecutter.broker_type == "kafka" %}confluent-kafka>=2.1.1{% elif cookiecutter.broker_type == "rabbitmq" %}aio-pika>=9.0.0{% else %}redis>=4.5.0{% endif %}

# Configuration and validation
{% if cookiecutter.use_pydantic == "yes" %}pydantic>=2.0.0
pydantic-settings>=2.0.0{% endif %}

# Logging
loguru>=0.7.0

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# ruff>=0.1.0
